Remove specific validation and move common joi to index, resolve circular
